requirements.txt
# ── Core dependencies (Python 3.9 / 3.10) ──────────────────────────
PySide6==6.7.2
PySide6-Addons==6.7.2        # QMediaPlayer 等附加组件

# yt-dlp 是下载模块的核心，必须用最新版
yt-dlp==2024.9.post0

# Faster-Whisper：纯 CPU 推理（device="cpu", compute_type="int8"）
faster-whisper==1.0.3

# FFmpeg Python 封装，内部会调用 ffmpeg.exe
ffmpeg-python==0.2.0

# 纯 Python 依赖（无 C 扩展冲突）
huggingface_hub>=0.20.0
tokenizers>=0.15.0           # Faster-Whisper 间接依赖

# ── PyInstaller 打包用（开发时可选） ────────────────────────────
# pyinstaller==6.10.0
# PyInstaller-hooks-contrib>=2024.0
